requirements.txt
Flask==2.3.2
opencv-python==4.7.0.72
numpy==1.24.3
pandas==2.0.1
python-dotenv==1.0.0
boto3==1.26.152
botocore==1.29.152
